Implementação básica do riverpod com a primeira separação de arquivos e protótipo no uso do provider. Pequeno ajuste no teste para comportar a mudança de nome da classe
