+++ lib/app.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import 'home_page.dart';
+
+final riverpodPrototypeProvider = Provider((_) => 'Testing Riverpod Provider');
+
+class App extends ConsumerWidget {
+  const App({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final String value = ref.watch(riverpodPrototypeProvider);
+
+    return MaterialApp(
+      title: 'Ricky and Morty',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: HomePage(title: value),
+    );
+  }
+}
